Split Owner and OwnerSearch models
The OwnerSearch is not a model, it's just a namespace,
so make it a real class to be able to pass state
