+++ src/api/app/controllers/webui/search_controller.rb
@@ -17,7 +17,7 @@ class Webui::SearchController < Webui::WebuiController
       return
     end
 
-    @results = Owner.search({ limit: @owner_limit.to_s, devel: @owner_devel.to_s }, @search_text)
+    @results = OwnerSearch.new({ limit: @owner_limit.to_s, devel: @owner_devel.to_s }).for(@search_text)
     flash[:notice] = 'Your search did not return any results.' if @results.empty?
   end
